requirements.txt
browser-cookie3
requests-html
PyQt5-sip
PyQt5
pyyaml
python-crontab
pyinstaller
pytest
requests-file
apscheduler
XlsxWriter
python-dateutil